+++ bin/impl/install/zookeeper.sh
@@ -28,12 +28,15 @@ verify_exist_hash "$ZOOKEEPER_TARBALL" "$ZOOKEEPER_HASH"
 
 print_to_console "Installing Apache ZooKeeper $ZOOKEEPER_VERSION at $ZOOKEEPER_HOME"
 
-rm -rf "$INSTALL"/*zookeeper-*
-rm -f "$ZOO_LOG_DIR"/*
-rm -rf "$DATA_DIR"/zookeeper
+rm -rf "${INSTALL:?}"/*zookeeper-*
+rm -f "${ZOO_LOG_DIR:?}"/*
+rm -rf "${DATA_DIR:?}"/zookeeper
 mkdir -p "$ZOO_LOG_DIR"
 
 tar xzf "$DOWNLOADS/$ZOOKEEPER_TARBALL" -C "$INSTALL"
 
 cp "$UNO_HOME"/conf/zookeeper/* "$ZOOKEEPER_HOME"/conf/
 $SED "s#DATA_DIR#$DATA_DIR#g" "$ZOOKEEPER_HOME"/conf/zoo.cfg
+
+true
+# zookeeper.sh
